Update Aruba, Cucumber, and its dependency.
This is the only way we can make Aruba play nice with Rake 0.9.0, as Aruba was locking 0.9.x version of Cucumber internally, and we want 0.10.x version of Cucumber which already has the Rake fix.

Note that the new Aruba 0.3.x adds the timeout algorithm, which will kills the process after 3 seconds. I've also have to patch that to wait for 5 minutes as Bundler is very slow during testing.