Ensure Bundler is available in vendored path
This fixes an issue where user would get "Could not find bundler" error
when tries to use Appraisal in project that vendorize gems.

The root cause was that Bundler will not bundle itself when you run
`bundle install` with `--path` option. As Appraisal resets `GEM_HOME`
to avoid dependency conflict, it also made RubyGems unable to locate
code for Bundler as well.

Fix #13, #30 and #87.