Make git_source work in nested groups
If you declare a `git_source` at the top level of your Gemfile or a spec in
Appraisals, it will substitute the gem source in `gem` declarations. If you
want to use those gem sources in nested groups declared with `group`,
`platforms`, `platform`, `source`, `git`, or `path`, the `git_source`
declarations from the top level aren't inherited, so they don't work like they
do with bundler.

Fortunately, in `BundlerDSL` there's a protected `git_sources` writer, and all
the groups are also declared in that file. It should be easy enough to assign
all of the group objects to a copy of the `git_sources`. That would make the
substitutions work.

Fixes #159.