Switch to using Bundler.with_original_env
This fixes a long-running issue with more recent Bundler versions and
compatibility with GitHub Actions (and other CI services), where the
environment wasn't being handled correctly.

Introduce a new 'APPRAISAL_UNDER_TEST' environment variable that
`Command` uses to determine if Appraisal itself is under test, and if
so, tweaks the environment in the way the tests require. This allows the
tests to pass while using `Bundler.with_original_env`.

Fixes #173