Add support for `ruby file:` Gemfile syntax
Previously, this would yield an error:

    appraise "foo" do
      ruby file: '../.ruby-version'

      gem "faraday-follow_redirects", "0.3.0"
    end

`Appraisal::BundlerDSL#ruby_version` is emitting a block when the
Gemfile DSL ruby method is used. Currently, this is emitted, which is
not a method argument but rather a block:

    ruby {:file=>"../.ruby-version"}

To fix this, we need to add `()` to ensure the arguments stay arguments:

    ruby({:file=>"../.ruby-version"})

This has been tested on Bundler 2.3.27 only — at this point in time we
have compatibility issues with others.

https://github.com/thoughtbot/appraisal/issues/218